% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{alerts}
\alias{alerts}
\title{Weather Alerts for United States and Europe}
\usage{
alerts(location, key = get_api_key(), raw = FALSE, raw_JSON = FALSE,
  message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{raw_JSON}{if TRUE return entire alert as JSON}

\item{message}{if TRUE print out requested URL}
}
\value{
A string containing alert type, message, start time and expiration.
}
\description{
Weather Alerts for United States and Europe
}
\examples{
\dontrun{
alerts(set_location(territory = "Hawaii", city = "Honolulu"))
alerts(set_location(airport_code = "SEA"))
alerts(set_location(zip_code = "90210"))
alerts(set_location(territory = "IR", city = "Tehran"))
}
}
