% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{as.numeric.nonempty}
\alias{as.numeric.nonempty}
\title{as.numeric with special handling for length 0 (NULL) objects}
\usage{
\method{as.numeric}{nonempty}(x)
}
\arguments{
\item{x}{the object to cast as numeric}
}
\value{
value of type double
}
\description{
as.numeric with special handling for length 0 (NULL) objects
}
