% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomy.R
\name{astronomy}
\alias{astronomy}
\title{Moon phase, sunrise and sunset times for today.}
\usage{
astronomy(location, key = get_api_key(), raw = FALSE, message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
tbl_df with: location, moon phase, percent visible,
        moon rise and set times, sun rise and set times.
}
\description{
Moon phase, sunrise and sunset times for today.
}
\examples{
\dontrun{
astronomy(set_location(territory = "Hawaii", city = "Honolulu"))
astronomy(set_location(airport_code = "SEA"))
astronomy(set_location(zip_code = "90210"))
astronomy(set_location(territory = "IR", city = "Tehran"))
}
}
