% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R, R/url.R
\name{base_url}
\alias{base_url}
\title{Base URL for wunderground API}
\usage{
base_url()

base_url()
}
\value{
base wunderground URL

base wunderground URL
}
\description{
Base URL for wunderground API

Base URL for wunderground API
}

