% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{conditions}
\alias{conditions}
\title{Current conditions including current temperature, weather condition,
humidity, wind, feels-like, temperature, barometric pressure,
and visibility.}
\usage{
conditions(location, use_metric = FALSE, key = get_api_key(), raw = FALSE,
  message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{use_metric}{Metric or imperial units}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
tbl_df with conditions
}
\description{
Current conditions including current temperature, weather condition,
humidity, wind, feels-like, temperature, barometric pressure,
and visibility.
}
\examples{
\dontrun{
conditions(set_location(territory = "Hawaii", city = "Honolulu"))
conditions(set_location(airport_code = "SEA"))
conditions(set_location(zip_code = "90210"))
conditions(set_location(territory = "IR", city = "Tehran"))
}
}
