% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane.R
\name{current_hurricane}
\alias{current_hurricane}
\title{Current hurricane - within the US only.
Note: all times in eastern}
\usage{
current_hurricane(key = get_api_key(), use_metric = FALSE, raw = FALSE,
  message = TRUE)
}
\arguments{
\item{key}{weather underground API key}

\item{use_metric}{Metric or imperial units}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
Hurricane info
}
\description{
Current hurricane - within the US only.
Note: all times in eastern
}
\examples{
\dontrun{
current_hurricane()
}
}
