% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{dst_repeat_starttime}
\alias{dst_repeat_starttime}
\title{Given a tz name, find the bounds of the interval where hh:mm does not 
  uniquely determine an instant in time without also knowing whether 
  DST is in effect.  Assumes that DST transitions happen on hour
  boundaries, which is true almost everywhere, and that the wall clock 
  shifts back and repeats exactly 1 hour.
  This relies on R and the OS to properly manage DST in all timezones.}
\usage{
dst_repeat_starttime(y, m, d, tz)
}
\arguments{
\item{y}{the year}

\item{m}{the month}

\item{d}{the day}

\item{tz}{the timezone}
}
\value{
list of two integer betweeen 0000 and 2359, hhmm format. 
  the first integer is the beginning of the interval of clock times which
  correspond to 2 separate instants of time, the second is the end of that
}
\description{
Given a tz name, find the bounds of the interval where hh:mm does not 
  uniquely determine an instant in time without also knowing whether 
  DST is in effect.  Assumes that DST transitions happen on hour
  boundaries, which is true almost everywhere, and that the wall clock 
  shifts back and repeats exactly 1 hour.
  This relies on R and the OS to properly manage DST in all timezones.
}
