% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{encode_NA}
\alias{encode_NA}
\title{Processes data.frames and replaces wunderground's -9999/-999 to NAs}
\usage{
encode_NA(df)
}
\arguments{
\item{data.frame}{the data.frame to process}
}
\value{
data.frame with correctly encoded NAs
}
\description{
Processes data.frames and replaces wunderground's -9999/-999 to NAs
}

