% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast3day}
\alias{forecast3day}
\title{Forecast for the next 3 days.}
\usage{
forecast3day(location, use_metric = FALSE, key = get_api_key(),
  raw = FALSE, message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{use_metric}{Metric or imperial units}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
tbl_df with date (in posix format), high and low temp,
        conditions, precipitation, rain, snow,
        max and avg wind speed, max/min and avg humidity
}
\description{
Forecast for the next 3 days.
}

