% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geolookup.R
\name{geolookup}
\alias{geolookup}
\title{Lists nearby weather stations for a given location}
\usage{
geolookup(location, use_metric = TRUE, key = get_api_key(), raw = FALSE,
  message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{use_metric}{Metric or imperial units}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
tbl_df of nearby weather stations with:
        type, city, state, country, id, lat, lon and
        dist (in either mi or km)
}
\description{
Lists nearby weather stations for a given location
}

