% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/key.R
\name{has_api_key}
\alias{has_api_key}
\title{Detects if wunderground API key is set}
\usage{
has_api_key()
}
\value{
TRUE if API key set, otherwise FALSE
}
\description{
Detects if wunderground API key is set
}

