% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/history.R
\name{history_daily}
\alias{history_daily}
\title{Summarized weather data for specified date.}
\usage{
history_daily(location, date = "20150101", date_fmt = "pretty",
  use_metric = TRUE, key = get_api_key(), raw = FALSE, message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{date}{Date as YYYYMMDD format}

\item{date_fmt}{date format to return}

\item{use_metric}{Metric or imperial units}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}

\item{col_names}{date format to return}
}
\value{
tbl_df of summarized weather
}
\description{
Summarized weather data for specified date.
}

