% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{is_fall_back_day}
\alias{is_fall_back_day}
\title{Check if Year, Month, Day, TZ is a transition from Daylight to Standard Time}
\usage{
is_fall_back_day(y, m, d, tz)
}
\arguments{
\item{y}{the year}

\item{m}{the month}

\item{d}{the day}

\item{tz}{the timezone}
}
\value{
logical
}
\description{
Check if Year, Month, Day, TZ is a transition from Daylight to Standard Time
}
