% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{is_valid_airport}
\alias{is_valid_airport}
\title{Checks if airport code is valid}
\usage{
is_valid_airport(name)
}
\arguments{
\item{name}{Airport code either IATA or ICAO}
}
\value{
TRUE if valid otherwise FALSE
}
\description{
Checks if airport code is valid
}
