% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{is_valid_territory}
\alias{is_valid_territory}
\title{Checks if country/state is a valid one}
\usage{
is_valid_territory(name)
}
\arguments{
\item{name}{Name of state or country}
}
\value{
TRUE if valid state or country otherwise FALSE
}
\description{
Checks if country/state is a valid one
}
