% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/location.R
\name{list_airports}
\alias{list_airports}
\title{Returns a data.frame of valid airport codes (ICAO and IATA).}
\usage{
list_airports()
}
\value{
data.frame of airport codes with country and city
}
\description{
Returns a data.frame of valid airport codes (ICAO and IATA).
}

