% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/location.R
\name{list_countries}
\alias{list_countries}
\title{Returns a data.frame of valid countries with iso abbreviations and region}
\usage{
list_countries()
}
\value{
data.frame of valid country names with iso codes
}
\description{
Returns a data.frame of valid countries with iso abbreviations and region
}
\examples{
list_countries()
}

