% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{list_states}
\alias{list_states}
\title{Returns a data.frame of valid states with abbreviations and regions}
\usage{
list_states()
}
\value{
data.frame of states with abbreviation and region
}
\description{
Returns a data.frame of valid states with abbreviations and regions
}
\examples{
\dontrun{
list_states()
}
}
