% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{lookup_country_code}
\alias{lookup_country_code}
\title{Lookup ISO country code
weatherunderground API doesn't recognize iso codes uniformly for every country.name}
\usage{
lookup_country_code(name, region = NULL)
}
\arguments{
\item{name}{Name of country}

\item{region}{Geographic region}
}
\value{
data.frame of country codes
}
\description{
Lookup ISO country code
weatherunderground API doesn't recognize iso codes uniformly for every country.name
}
\examples{
\dontrun{
lookup_country_code("Korea")
lookup_country_code("Guinea", region = "Africa")
}
}
