% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{measurement_exists}
\alias{measurement_exists}
\title{Check if a variable exists for a PWS. If not set the value to -9999}
\usage{
measurement_exists(x, class = "numeric")
}
\arguments{
\item{x}{the value to check}

\item{class}{a character given the desired class for the variable}
}
\description{
Check if a variable exists for a PWS. If not set the value to -9999
}
