% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{nonempty}
\alias{nonempty}
\title{return object, or NA for length 0 (NULL) objects}
\usage{
nonempty(x)
}
\arguments{
\item{x}{the object to cast as numeric}
}
\value{
value of type double
}
\description{
return object, or NA for length 0 (NULL) objects
}
