\name{rwunderground-package}
\alias{rwunderground-package}
\alias{rwunderground}
\docType{package}
\title{
\packageTitle{rwunderground}
}
\description{
\packageDescription{rwunderground}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rwunderground}
\packageIndices{rwunderground}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rwunderground}

Maintainer: \packageMaintainer{rwunderground}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
