% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.R
\name{satellite}
\alias{satellite}
\title{Returns image URL for satellite imagery}
\usage{
satellite(location, key = get_api_key(), raw = FALSE, message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
URL to satellite imagery
}
\description{
Returns image URL for satellite imagery
}
\examples{
\dontrun{
satellite(set_location(territory = "Hawaii", city = "Honolulu"))
satellite(set_location(territory = "Washington", city = "Seattle"))
satellite(set_location(territory = "Louisiana", city = "New Orleans"))
}
}
