% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_api_key}
\alias{set_api_key}
\title{Sets the wunderground API key}
\usage{
set_api_key(key)
}
\arguments{
\item{key}{wunderground API key}
}
\value{
API key
}
\description{
Sets the wunderground API key
}
\examples{
\dontrun{
set_api_key("1a2b3c4d")
}
}
