% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/location.R
\name{set_location}
\alias{set_location}
\title{Specifies location of request}
\usage{
set_location(zip_code = NULL, territory = NULL, city = NULL,
  airport_code = NULL, PWS_id = NULL, lat_long = NULL, autoip = NULL)
}
\arguments{
\item{zip_code}{}

\item{territory}{}

\item{city}{}

\item{airport_code}{}

\item{PWS_id}{}

\item{lat_long}{}

\item{autoip}{}
}
\value{
formatted and validated location string
}
\description{
This is a wrapper function that will validate and format location strings
for requesting data from weather underground.
}

