% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{stop_for_error}
\alias{stop_for_error}
\title{Detect and stop for any wunderground request errors}
\usage{
stop_for_error(httr_parsed_req)
}
\arguments{
\item{httr_parsed_req}{httr request object}
}
\description{
Detect and stop for any wunderground request errors
}
