% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide.R
\name{tide}
\alias{tide}
\title{Tidal information for a location within the USA.
Tidal information only available for US cities.  Units are in feet.}
\usage{
tide(location, key = get_api_key(), raw = FALSE, message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
tbl_df with date, height and type
}
\description{
Tidal information for a location within the USA.
Tidal information only available for US cities.  Units are in feet.
}
\examples{
\dontrun{
tide(set_location(territory = "Hawaii", city = "Honolulu"))
tide(set_location(territory = "Washington", city = "Seattle"))
tide(set_location(territory = "Louisiana", city = "New Orleans"))
}
}
