% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac.R
\name{almanac}
\alias{almanac}
\title{Average and record high and low temperatures for current date going back
as far as weather underground has data or from the national weather service
going back 30 years.}
\usage{
almanac(location, use_metric = FALSE, key = get_api_key(), raw = FALSE,
  message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{use_metric}{Metric or imperial units}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}
}
\value{
tbl_df with columns: location, airport,
        avg_high, record high, avg_low, record low.
}
\description{
Average and record high and low temperatures for current date going back
as far as weather underground has data or from the national weather service
going back 30 years.
}
\examples{
\dontrun{
almanac(set_location(territory = "Hawaii", city = "Honolulu"))
almanac(set_location(airport_code = "SEA"))
almanac(set_location(zip_code = "90210"))
almanac(set_location(territory = "IR", city = "Tehran"))
}
}
