% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{dst_POSIXct}
\alias{dst_POSIXct}
\title{Return POSIXct time from 7 variables}
\usage{
dst_POSIXct(y, m, d, hr, mn, sec, tz)
}
\arguments{
\item{y}{vector of years}

\item{m}{vector of months}

\item{d}{vector of days}

\item{hr}{vector of hours}

\item{mn}{vector of minutes}

\item{sec}{vector of seconds}

\item{tz}{vector of timezones}
}
\value{
POSIXct time assuming vectors sorted by true chronological order,
    at least for the hour that "occurs twice", once with Daylight Time, 
    then again with Standard Time.  If there are no nonmonotonicities in 
    the times, all times in this hour will be assumed to be Daylight Time
}
\description{
Return POSIXct time from 7 variables
}
