% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{dst_repeat_starttime}
\alias{dst_repeat_starttime}
\title{Find the text to POSIXct ambiguous interval.}
\usage{
dst_repeat_starttime(y, m, d, tz)
}
\arguments{
\item{y}{the year}

\item{m}{the month}

\item{d}{the day}

\item{tz}{the timezone}
}
\value{
list of two integers betweeen 0000 and 2359, hhmm format. 
  the first integer is the beginning of the interval of clock times which
  correspond to 2 separate instants of time, the second is the end of that
  interval.  The left endpoint is ambiguous, the right endpoint is not
  since it maps only to Standard Time.
}
\description{
Assumes that DST transitions happen on hour boundaries, which is true
   almost everywhere, and that the wall clock shifts back and repeats 
   exactly 1 hour, again true almost everywhere. This code relies on R 
   and the OS to properly manage DST in all timezones.
}
