% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/history.R
\name{history_range}
\alias{history_range}
\title{Hourly weather data for specified date range.}
\usage{
history_range(location, date_start = "20150101", date_end = "20150105",
  limit = FALSE, no_api = FALSE, date_fmt = "pretty", use_metric = TRUE,
  key = get_api_key(), raw = FALSE, message = TRUE)
}
\arguments{
\item{location}{location set by set_location}

\item{date_fmt}{date format to return}

\item{use_metric}{Metric or imperial units}

\item{key}{weather underground API key}

\item{raw}{if TRUE return raw httr object}

\item{message}{if TRUE print out requested URL}

\item{date}{Date as YYYYMMDD format}

\item{col_names}{date format to return}
}
\value{
tbl_df with date, temperature, dew point,
        humidity, wind speed, gust and direction,
        visibility, pressure, wind chill, heat index,
        precipitation, condition, fog, rain, snow,
        hail, thunder, tornado
}
\description{
Hourly weather data for specified date range.
}

