% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{list_airports}
\alias{list_airports}
\title{Returns a data.frame of valid airport codes (ICAO and IATA).}
\usage{
list_airports()
}
\value{
data.frame of airport codes with country and city
}
\description{
This dataset is from the openflights.org airport database.  It can be
found at \url{http://openflights.org/data.html#airport}.  This data
is provided under the open database license -- more information can
be found here: \url{http://opendatacommons.org/licenses/odbl/1.0/}.
}
\examples{
\dontrun{
list_airports()
}
}
