% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/planner.R
\name{planner}
\alias{planner}
\title{}
\usage{
planner(location, date_fmt = "pretty", use_metric = TRUE,
  start_date = "0501", end_date = "0531", key = get_api_key(),
  raw = FALSE, message = TRUE)
}
\description{

}

